Extend tests
